TaskManager/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.3.4=pypi_0
attrs=21.2.0=pypi_0
autopep8=1.5.7=pyhd3eb1b0_0
blinker=1.4=pypi_0
ca-certificates=2021.4.13=haa95532_1
certifi=2020.12.5=py38haa95532_0
chardet=3.0.4=pypi_0
click=6.7=pypi_0
colorama=0.4.4=pypi_0
coreapi=2.3.3=pypi_0
coreschema=0.0.4=pypi_0
coverage=5.1=pypi_0
dj-rest-auth=2.1.5=pypi_0
django=3.1.8=pypi_0
django-cors-headers=3.3.0=pypi_0
django-extensions=2.2.9=pypi_0
django-filter=2.4.0=pypi_0
django-nose=1.4.6=pypi_0
django-rest-auth=0.9.5=pypi_0
djangorestframework=3.11.2=pypi_0
djangorestframework-simplejwt=4.4.0=pypi_0
drf-util=0.0.12=pypi_0
drf-yasg=1.17.1=pypi_0
elasticsearch=6.3.1=pypi_0
faker=4.1.0=pypi_0
flex=6.13.2=pypi_0
httpie=2.4.0=pypi_0
idna=2.7=pypi_0
inflection=0.5.1=pypi_0
itypes=1.2.0=pypi_0
jsonpointer=1.14=pypi_0
jsonschema=3.2.0=pypi_0
nose=1.3.7=pypi_0
openssl=1.1.1k=h2bbff1b_0
packaging=20.9=pypi_0
pip=21.0.1=py38haa95532_0
pycodestyle=2.7.0=pyhd3eb1b0_0
pygments=2.9.0=pypi_0
pyjwt=1.7.1=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.17.3=pypi_0
pysocks=1.7.1=pypi_0
python=3.8.8=hdbf39b2_5
python-dateutil=2.8.1=pypi_0
pytz=2021.1=pypi_0
pyyaml=3.13=pypi_0
requests=2.25.1=pypi_0
requests-toolbelt=0.9.1=pypi_0
rfc3987=1.3.8=pypi_0
ruamel-yaml=0.17.4=pypi_0
ruamel-yaml-clib=0.2.2=pypi_0
setuptools=52.0.0=py38haa95532_0
sqlite=3.35.4=h2bbff1b_0
sqlparse=0.4.1=pypi_0
strict-rfc3339=0.7=pypi_0
swagger-spec-validator=2.4.1=pypi_0
text-unidecode=1.3=pypi_0
toml=0.10.2=pyhd3eb1b0_0
uritemplate=3.0.1=pypi_0
urllib3=1.24.3=pypi_0
validate-email=1.3=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py38_0
psycopg2-binary>=2.8
